requirements.txt
absl-py
pylint
yapf
numpy
qmsolve[with_mayavi]
pydub
streamlit
st-pages